+++ src/main/java/com/alibaba/middleware/race/sync/util/RecordMap.java
@@ -29,11 +29,11 @@ public class RecordMap {
 
 	private byte[]			data;
 
-	private int[]				powers;
+	private byte[]			powers;
 
 	private AtomicIntegerArray	lock;
 
-	private int[]				versions;
+	private short[]			versions;
 
 	private int				capacity;
 
@@ -49,8 +49,8 @@ public class RecordMap {
 	private void init(int capacity, int cols) {
 		this.recordLen = 8 * cols;
 		this.data = new byte[capacity * recordLen];
-		this.versions = new int[cols * capacity];
-		this.powers = new int[capacity];
+		this.versions = new short[cols * capacity];
+		this.powers = new byte[capacity];
 		this.lock = new AtomicIntegerArray(capacity);
 	}
 
@@ -105,7 +105,7 @@ public class RecordMap {
 		lock.set(idx, 0);
 	}
 
-	public void setColumn(int pk, byte name, int version, byte[] src, int off, int len) {
+	public void setColumn(int pk, byte name, short version, byte[] src, int off, int len) {
 		int iPk = ix(pk);
 		int vI = iPk * cols + name;
 		if (version < versions[vI]) {
